Restructure existing form requests
