--- app/Http/Requests/Api/V1/Auth/SocialLoginRequest.php
@@ -1,29 +1,0 @@
-<?php
-
-namespace App\Http\Requests\Api\V1\Auth;
-
-use Illuminate\Foundation\Http\FormRequest;
-
-class SocialLoginRequest extends FormRequest
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     */
-    public function authorize(): bool
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array|string>
-     */
-    public function rules(): array
-    {
-        return [
-            'provider' => ['required', 'string', 'in:google,telegram,github'], // Add other providers as needed
-            'access_token' => ['required', 'string'],
-        ];
-    }
-}
