+++ app/Http/Requests/Auth/VerifyRequest.php
@@ -1,28 +1,27 @@
 <?php
 
-namespace App\Http\Requests\Api\V1\Auth;
+namespace App\Http\Requests\Auth;
 
 use Illuminate\Foundation\Http\FormRequest;
 
 class VerifyRequest extends FormRequest
 {
-    /**
-     * Determine if the user is authorized to make this request.
-     */
     public function authorize(): bool
     {
         return true;
     }
 
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array|string>
-     */
     public function rules(): array
     {
         return [
             'token' => ['required', 'string', 'exists:registrations,verification_token'],
+        ];
+    }
+
+    public function messages(): array
+    {
+        return [
+            'token.exists' => 'Invalid or expired verification token',
         ];
     }
 }
